Fix code review findings in GraphDataStore and CommunicationStatsModel
- Add missing <algorithm> include in graphdatastore.cpp (std::sort was
  relying on transitive inclusion)
- Fix qint16 local variable in convertToActiveGraphIndex to match the
  qint32 return type
- Remove no-op if guards in setStartTime/setEndTime (no signal is
  emitted, so the guards served no purpose)
- Add missing Doxygen comments to all undocumented public methods in
  CommunicationStatsModel

+++ src/models/graphdatastore.cpp
@@ -3,6 +3,8 @@
 
 #include "util/util.h"
 
+#include <algorithm>
+
 GraphDataStore::GraphDataStore(QObject* parent) : QObject(parent), _selectedGraphIdx(-1)
 {
 }
@@ -191,7 +193,7 @@ void GraphDataStore::activeGraphIndexList(QList<quint16>* pList) const
 
 qint32 GraphDataStore::convertToActiveGraphIndex(quint32 graphIdx) const
 {
-    qint16 result = -1;
+    qint32 result = -1;
     for (qint32 activeIdx = 0; activeIdx < _activeGraphList.size(); activeIdx++)
     {
         if (_activeGraphList[activeIdx] == graphIdx)
